requirements.txt
click==6.6
Flask==0.11.1
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
Werkzeug==0.11.10
wheel==0.24.0
boto3==1.3.1
antlr4-python2-runtime==4.5.3